ISIS-3010: metamodel bug fixes; restores CommandLog UI for JDO

+++ core/metamodel/src/main/java/org/apache/isis/core/metamodel/util/Facets.java
@@ -321,7 +321,7 @@ public final class Facets {
         final Integer result = (typicalLength != null
                 && maxLength.isPresent()
                 && typicalLength > maxLength.getAsInt())
-                ? maxLength.getAsInt()
+                ? (Integer)maxLength.getAsInt()
                 : typicalLength;
         return Optional.ofNullable(result)
                 .map(OptionalInt::of)
